Fixed major bug after changing the lsns from int to long
git-svn-id: https://hyracks.googlecode.com/svn/branches/hyracks_btree_updates_next@699 123451ca-8445-de46-9d55-352943316053

+++ hyracks-storage-am-rtree/src/main/java/edu/uci/ics/hyracks/storage/am/rtree/frames/RTreeNSMFrame.java
@@ -86,7 +86,7 @@ public abstract class RTreeNSMFrame extends TreeIndexNSMFrame implements
 	@Override
 	public void initBuffer(byte level) {
 		super.initBuffer(level);
-		buf.putInt(pageNsnOff, 0);
+		buf.putLong(pageNsnOff, 0);
 		buf.putInt(rightPageOff, -1);
 	}
